Move service tests and update AvailApi test.
